Update README.md
The data has been taken for the data professional salaries for different countries for the year 2022 in the month of June. After carefully digging deep into the data, the data has been transformed, and decided to display the key insights from the data by using different charts.

The first tree map which represents the country of survey takers, the main countries from which the data has been collected are United states, India, United Kingdom, Canada and others, when you will click on each country present in the treemap, it will display the number of data professional living in a respective country.

The second column chart which represents the favourite programming languages by different data professionals.

There are two cards at the top right which displaying the number of people who took the survey and the other card visual displaying the average age of the survey taker.

The two Gauge visuals displaying the happiness with work-life balance and the happiness with the salary.

The stacked bar chart which is displaying the average salary by job title in the descending order.

The donut chart which represent how difficult it was for data professional to break into the data.